move account select into active account component
