+++ deployables/extension/src/panel/pages/$activeAccountId/ActiveAccount.tsx
@@ -1,24 +1,42 @@
-import { getAccount, getActiveRoute, saveActiveAccount } from '@/accounts'
-import { ProvideAccount } from '@/companion'
+import {
+  getAccount,
+  getAccounts,
+  getActiveRoute,
+  saveActiveAccount,
+} from '@/accounts'
+import { ProvideAccount, useCompanionAppUrl } from '@/companion'
 import { ProvideExecutionRoute } from '@/execution-routes'
 import { ProvideProvider } from '@/providers-ui'
 import { sentry } from '@/sentry'
-import { getActiveTab, sendMessageToCompanionApp } from '@/utils'
+import {
+  getActiveTab,
+  getInt,
+  getString,
+  sendMessageToCompanionApp,
+} from '@/utils'
+import { getCompanionAppUrl } from '@zodiac/env'
 import {
   CompanionAppMessageType,
   CompanionResponseMessageType,
   useTabMessageHandler,
 } from '@zodiac/messages'
+import { encode } from '@zodiac/schema'
+import { GhostLinkButton, Page } from '@zodiac/ui'
+import { ArrowUpFromLine, Landmark } from 'lucide-react'
 import {
   Outlet,
   redirect,
   useLoaderData,
+  type ActionFunctionArgs,
   type LoaderFunctionArgs,
 } from 'react-router'
+import { AccountSelect } from './AccountSelect'
 import { getActiveAccountId } from './getActiveAccountId'
+import { Intent } from './intents'
 
 export const loader = async ({ params, request }: LoaderFunctionArgs) => {
   const activeAccountId = getActiveAccountId(params)
+  const accounts = await getAccounts({ signal: request.signal })
 
   try {
     const [account, route] = await Promise.all([
@@ -44,6 +62,7 @@ export const loader = async ({ params, request }: LoaderFunctionArgs) => {
     return {
       route,
       account,
+      accounts,
     }
   } catch (error) {
     await saveActiveAccount(null)
@@ -54,8 +73,91 @@ export const loader = async ({ params, request }: LoaderFunctionArgs) => {
   }
 }
 
+export const action = async ({ request, params }: ActionFunctionArgs) => {
+  const data = await request.formData()
+
+  switch (getString(data, 'intent')) {
+    case Intent.EditAccount: {
+      const windowId = getInt(data, 'windowId')
+      const tabs = await chrome.tabs.query({ windowId })
+
+      const accountId = getString(data, 'accountId')
+      const account = await getAccount(accountId, { signal: request.signal })
+
+      if (account.remote) {
+        const existingTab = tabs.find(
+          (tab) =>
+            tab.url != null &&
+            tab.url.startsWith(`${getCompanionAppUrl()}/account/${accountId}`),
+        )
+
+        if (existingTab != null && existingTab.id != null) {
+          await chrome.tabs.update(existingTab.id, {
+            active: true,
+            url: `${getCompanionAppUrl()}/account/${accountId}`,
+          })
+        } else {
+          await chrome.tabs.create({
+            active: true,
+            url: `${getCompanionAppUrl()}/account/${accountId}`,
+          })
+        }
+      } else {
+        const route = await getActiveRoute(accountId)
+
+        const existingTab = tabs.find(
+          (tab) =>
+            tab.url != null &&
+            tab.url.startsWith(`${getCompanionAppUrl()}/edit/${accountId}`),
+        )
+
+        if (existingTab != null && existingTab.id != null) {
+          await chrome.tabs.update(existingTab.id, {
+            active: true,
+            url: `${getCompanionAppUrl()}/edit/${accountId}/${encode(route)}`,
+          })
+        } else {
+          await chrome.tabs.create({
+            active: true,
+            url: `${getCompanionAppUrl()}/edit/${accountId}/${encode(route)}`,
+          })
+        }
+      }
+
+      return null
+    }
+
+    case Intent.ListAccounts: {
+      const windowId = getInt(data, 'windowId')
+      const tabs = await chrome.tabs.query({ windowId })
+
+      const existingTab = tabs.find(
+        (tab) => tab.url != null && tab.url === `${getCompanionAppUrl()}/edit`,
+      )
+
+      if (existingTab != null && existingTab.id != null) {
+        await chrome.tabs.update(existingTab.id, { active: true })
+      } else {
+        await chrome.tabs.create({
+          active: true,
+          url: `${getCompanionAppUrl()}/edit`,
+        })
+      }
+
+      return null
+    }
+
+    case Intent.ActivateAccount: {
+      const accountId = getString(data, 'accountId')
+      const activeAccountId = getActiveAccountId(params)
+
+      return redirect(`/${activeAccountId}/clear-transactions/${accountId}`)
+    }
+  }
+}
+
 const ActiveRoute = () => {
-  const { route, account } = useLoaderData<typeof loader>()
+  const { route, account, accounts } = useLoaderData<typeof loader>()
 
   useTabMessageHandler(
     CompanionAppMessageType.REQUEST_ACTIVE_ROUTE,
@@ -71,7 +173,37 @@ const ActiveRoute = () => {
     <ProvideAccount account={account}>
       <ProvideExecutionRoute route={route}>
         <ProvideProvider>
-          <Outlet />
+          <Page>
+            <Page.Header>
+              <div className="my-2">
+                <AccountSelect accounts={accounts} />
+              </div>
+            </Page.Header>
+
+            <div className="flex p-2">
+              <GhostLinkButton
+                fluid
+                openInNewWindow
+                size="small"
+                icon={ArrowUpFromLine}
+                to={`${useCompanionAppUrl()}/tokens/send`}
+              >
+                Send tokens
+              </GhostLinkButton>
+
+              <GhostLinkButton
+                fluid
+                openInNewWindow
+                size="small"
+                icon={Landmark}
+                to={`${useCompanionAppUrl()}/tokens/balances`}
+              >
+                View balances
+              </GhostLinkButton>
+            </div>
+
+            <Outlet />
+          </Page>
         </ProvideProvider>
       </ProvideExecutionRoute>
     </ProvideAccount>
