+++ deployables/extension/src/panel/pages/$activeAccountId/intents.ts
@@ -1,3 +1,6 @@
 export enum Intent {
   disconnectProvider = 'disconnectProvider',
+  EditAccount = 'EditAccount',
+  ListAccounts = 'ListAccounts',
+  ActivateAccount = 'ActivateAccount',
 }
