+++ deployables/extension/src/panel/pages/$activeAccountId/transactions/intents.ts
@@ -1,6 +1,4 @@
 export enum Intent {
-  EditAccount = 'EditAccount',
-  ListAccounts = 'ListAccounts',
-  ActivateAccount = 'ActivateAccount',
   Login = 'Login',
+  EditAccount = 'EditAccount',
 }
